personalProjects/jobTonic/requirements.txt
flask==2.3.3
openai
python-dotenv
selenium
bs4
webdriver_manager